// src/generated/graphql.tsx
export type Maybe<T> = T | null;
export type Exact<T extends { [key: string]: unknown }> = { [K in keyof T]: T[K] };
export type MakeOptional<T, K extends keyof T> = Omit<T, K> & { [SubKey in K]?: Maybe<T[SubKey]> };
export type MakeMaybe<T, K extends keyof T> = Omit<T, K> & { [SubKey in K]: Maybe<T[SubKey]> };
/** All built-in and custom scalars, mapped to their actual values */
export type Scalars = {
  ID: string;
  String: string;
  Boolean: boolean;
  Int: number;
  Float: number;
  /**
   * The MBID scalar represents MusicBrainz identifiers, which are
   * 36-character UUIDs.
   */
  MBID: any;
  /** Language code, optionally with country and encoding. */
  Locale: any;
  /** Year, month (optional), and day (optional) in YYYY-MM-DD format. */
  Date: any;
  /**
   * An [Interested Parties Information](https://musicbrainz.org/doc/IPI)
   * (IPI) code is an identifying number assigned by the CISAC database for musical
   * rights management.
   */
  IPI: any;
  /**
   * The [International Standard Name Identifier](https://musicbrainz.org/doc/ISNI)
   * (ISNI) is an ISO standard for uniquely identifying the public identities of
   * contributors to media content.
   */
  ISNI: any;
  /**
   * The [International Standard Recording Code](https://musicbrainz.org/doc/ISRC)
   * (ISRC) is an identification system for audio and music video recordings. It is
   * standarized by the [IFPI](http://www.ifpi.org/) in ISO 3901:2001 and used by
   * IFPI members to assign a unique identifier to every distinct sound recording
   * they release. An ISRC identifies a particular [sound recording](https://musicbrainz.org/doc/Recording),
   * not the song itself. Therefore, different recordings, edits, remixes and
   * remasters of the same song will each be assigned their own ISRC. However, note
   * that same recording should carry the same ISRC in all countries/territories.
   * Songs are identified by analogous [International Standard Musical Work Codes](https://musicbrainz.org/doc/ISWC)
   * (ISWCs).
   */
  ISRC: any;
  /** A length of time, in milliseconds. */
  Duration: any;
  /**
   * An [Amazon Standard Identification Number](https://musicbrainz.org/doc/ASIN)
   * (ASIN) is a 10-character alphanumeric unique identifier assigned by Amazon.com
   * and its partners for product identification within the Amazon organization.
   */
  ASIN: any;
  /**
   * [Disc ID](https://musicbrainz.org/doc/Disc_ID) is the code
   * number which MusicBrainz uses to link a physical CD to a [release](https://musicbrainz.org/doc/Release)
   * listing.
   *
   * A release may have any number of disc IDs, and a disc ID may be linked to
   * multiple releases. This is because disc ID calculation involves a hash of the
   * frame offsets of the CD tracks.
   *
   * Different pressing of a CD often have slightly different frame offsets, and
   * hence different disc IDs.
   *
   * Conversely, two different CDs may happen to have exactly the same set of frame
   * offsets and hence the same disc ID.
   */
  DiscID: any;
  /** A time of day, in 24-hour hh:mm notation. */
  Time: any;
  /** A web address. */
  URLString: any;
  /** Decimal degrees, used for latitude and longitude. */
  Degrees: any;
  /**
   * The [International Standard Musical Work Code](https://musicbrainz.org/doc/ISWC)
   * (ISWC) is an ISO standard similar to ISBNs for identifying musical works /
   * compositions.
   */
  ISWC: any;
};

/**
 * The query root, from which multiple types of MusicBrainz
 * requests can be made.
 */
export type Query = {
  __typename?: 'Query';
  /** Perform a lookup of a MusicBrainz entity by its MBID. */
  lookup?: Maybe<LookupQuery>;
  /** Browse all MusicBrainz entities directly linked to another entity. */
  browse?: Maybe<BrowseQuery>;
  /** Search for MusicBrainz entities using Lucene query syntax. */
  search?: Maybe<SearchQuery>;
  /** Fetches an object given its ID */
  node?: Maybe<Node>;
  /**
   * A query for data on [Last.fm](https://www.last.fm/) that is not connected
   * to any particular MusicBrainz entity. This field is provided by the
   * Last.fm extension.
   */
  lastFM?: Maybe<LastFmQuery>;
  spotify: SpotifyQuery;
};
// ...
export type Area = Node & Entity & {
  __typename?: 'Area';
  /** The ID of an object */
  id: Scalars['ID'];
  /** The MBID of the entity. */
  mbid: Scalars['MBID'];
  /** The official name of the entity. */
  name?: Maybe<Scalars['String']>;
  /**
   * The string to use for the purpose of ordering by name (for
   * example, by moving articles like ‘the’ to the end or a person’s last name to
   * the front).
   */
  sortName?: Maybe<Scalars['String']>;
  /** A comment used to help distinguish identically named entitites. */
  disambiguation?: Maybe<Scalars['String']>;
  /**
   * [Aliases](https://musicbrainz.org/doc/Aliases) are used to store
   * alternate names or misspellings.
   */
  aliases?: Maybe<Array<Maybe<Alias>>>;
  /**
   * [ISO 3166 codes](https://en.wikipedia.org/wiki/ISO_3166) are
   * the codes assigned by ISO to countries and subdivisions.
   */
  isoCodes?: Maybe<Array<Maybe<Scalars['String']>>>;
  /**
   * The type of area (country, city, etc. – see the [possible
   * values](https://musicbrainz.org/doc/Area)).
   */
  type?: Maybe<Scalars['String']>;
  /**
   * The MBID associated with the value of the `type`
   * field.
   */
  typeID?: Maybe<Scalars['MBID']>;
  /** A list of artists linked to this entity. */
  artists?: Maybe<ArtistConnection>;
  /** A list of events linked to this entity. */
  events?: Maybe<EventConnection>;
  /** A list of labels linked to this entity. */
  labels?: Maybe<LabelConnection>;
  /** A list of places linked to this entity. */
  places?: Maybe<PlaceConnection>;
  /** A list of releases linked to this entity. */
  releases?: Maybe<ReleaseConnection>;
  /** Relationships between this entity and other entitites. */
  relationships?: Maybe<Relationships>;
  /** A list of collections containing this entity. */
  collections?: Maybe<CollectionConnection>;
  /** A list of tags linked to this entity. */
  tags?: Maybe<TagConnection>;
  /**
   * Chart data available for this area on [Last.fm](https://www.last.fm/), if
   * the area represents a country with an [ISO 3166 code](https://en.wikipedia.org/wiki/ISO_3166).
   * This field is provided by the Last.fm extension.
   */
  lastFM?: Maybe<LastFmCountry>;
};
// ...
export type Node = {
  /** The id of the object. */
  id: Scalars['ID'];
};

/** An entity in the MusicBrainz schema. */
export type Entity = {
  /** The MBID of the entity. */
  mbid: Scalars['MBID'];
};

/**
 * [Aliases](https://musicbrainz.org/doc/Aliases) are variant names
 * that are mostly used as search help: if a search matches an entity’s alias, the
 * entity will be given as a result – even if the actual name wouldn’t be.
 */
export type Alias = {
  __typename?: 'Alias';
  /** The aliased name of the entity. */
  name?: Maybe<Scalars['String']>;
  /**
   * The string to use for the purpose of ordering by name (for
   * example, by moving articles like ‘the’ to the end or a person’s last name to
   * the front).
   */
  sortName?: Maybe<Scalars['String']>;
  /**
   * The locale (language and/or country) in which the alias is
   * used.
   */
  locale?: Maybe<Scalars['Locale']>;
  /**
   * Whether this is the main alias for the entity in the
   * specified locale (this could mean the most recent or the most common).
   */
  primary?: Maybe<Scalars['Boolean']>;
  /**
   * The type or purpose of the alias – whether it is a variant,
   * search hint, etc.
   */
  type?: Maybe<Scalars['String']>;
  /**
   * The MBID associated with the value of the `type`
   * field.
   */
  typeID?: Maybe<Scalars['MBID']>;
};
// ...
export type Artist = Node & Entity & {
  __typename?: 'Artist';
  /** The ID of an object */
  id: Scalars['ID'];
  /** The MBID of the entity. */
  mbid: Scalars['MBID'];
  /** The official name of the entity. */
  name?: Maybe<Scalars['String']>;
  /**
   * The string to use for the purpose of ordering by name (for
   * example, by moving articles like ‘the’ to the end or a person’s last name to
   * the front).
   */
  sortName?: Maybe<Scalars['String']>;
  /** A comment used to help distinguish identically named entitites. */
  disambiguation?: Maybe<Scalars['String']>;
  /**
   * [Aliases](https://musicbrainz.org/doc/Aliases) are used to store
   * alternate names or misspellings.
   */
  aliases?: Maybe<Array<Maybe<Alias>>>;
  /**
   * The country with which an artist is primarily identified. It
   * is often, but not always, its birth/formation country.
   */
  country?: Maybe<Scalars['String']>;
  /**
   * The area with which an artist is primarily identified. It
   * is often, but not always, its birth/formation country.
   */
  area?: Maybe<Area>;
  /**
   * The area in which an artist began their career (or where
   * they were born, if the artist is a person).
   */
  beginArea?: Maybe<Area>;
  /**
   * The area in which an artist ended their career (or where
   * they died, if the artist is a person).
   */
  endArea?: Maybe<Area>;
  /**
   * The begin and end dates of the entity’s existence. Its exact
   * meaning depends on the type of entity.
   */
  lifeSpan?: Maybe<LifeSpan>;
  /**
   * Whether a person or character identifies as male, female, or
   * neither. Groups do not have genders.
   */
  gender?: Maybe<Scalars['String']>;
  /**
   * The MBID associated with the value of the `gender`
   * field.
   */
  genderID?: Maybe<Scalars['MBID']>;
  /** Whether an artist is a person, a group, or something else. */
  type?: Maybe<Scalars['String']>;
  /**
   * The MBID associated with the value of the `type`
   * field.
   */
  typeID?: Maybe<Scalars['MBID']>;
  /**
   * List of [Interested Parties Information](https://musicbrainz.org/doc/IPI)
   * (IPI) codes for the artist.
   */
  ipis?: Maybe<Array<Maybe<Scalars['IPI']>>>;
  /**
   * List of [International Standard Name Identifier](https://musicbrainz.org/doc/ISNI)
   * (ISNI) codes for the artist.
   */
  isnis?: Maybe<Array<Maybe<Scalars['ISNI']>>>;
  /** A list of recordings linked to this entity. */
  recordings?: Maybe<RecordingConnection>;
  /** A list of releases linked to this entity. */
  releases?: Maybe<ReleaseConnection>;
  /** A list of release groups linked to this entity. */
  releaseGroups?: Maybe<ReleaseGroupConnection>;
  /** A list of works linked to this entity. */
  works?: Maybe<WorkConnection>;
  /** Relationships between this entity and other entitites. */
  relationships?: Maybe<Relationships>;
  /** A list of collections containing this entity. */
  collections?: Maybe<CollectionConnection>;
  /** The rating users have given to this entity. */
  rating?: Maybe<Rating>;
  /** A list of tags linked to this entity. */
  tags?: Maybe<TagConnection>;
  /**
   * Images of the artist from [fanart.tv](https://fanart.tv/).
   * This field is provided by the fanart.tv extension.
   */
  fanArt?: Maybe<FanArtArtist>;
  /**
   * Artist images found at MediaWiki URLs in the artist’s URL relationships.
   * Defaults to URL relationships with the type “image”.
   * This field is provided by the MediaWiki extension.
   */
  mediaWikiImages: Array<Maybe<MediaWikiImage>>;
  /**
   * Data about the artist from [TheAudioDB](http://www.theaudiodb.com/), a good
   * source of biographical information and images.
   * This field is provided by TheAudioDB extension.
   */
  theAudioDB?: Maybe<TheAudioDbArtist>;
  /** Information about the artist on Discogs. */
  discogs?: Maybe<DiscogsArtist>;
  /**
   * Data about the artist from [Last.fm](https://www.last.fm/), a good source
   * for measuring popularity via listener and play counts. This field is
   * provided by the Last.fm extension.
   */
  lastFM?: Maybe<LastFmArtist>;
  /** The artist’s entry on Spotify. */
  spotify?: Maybe<SpotifyArtist>;
};
// ...
export type LifeSpan = {
  __typename?: 'LifeSpan';
  /** The start date of the entity’s life span. */
  begin?: Maybe<Scalars['Date']>;
  /** The end date of the entity’s life span. */
  end?: Maybe<Scalars['Date']>;
  /** Whether or not the entity’s life span has ended. */
  ended?: Maybe<Scalars['Boolean']>;
};
// ...
export type Rating = {
  __typename?: 'Rating';
  /** The number of votes that have contributed to the rating. */
  voteCount: Scalars['Int'];
  /** The average rating value based on the aggregated votes. */
  value?: Maybe<Scalars['Float']>;
};
// ...
export type SearchQueryArtistsArgs = {
  query: Scalars['String'];
  after?: Maybe<Scalars['String']>;
  first?: Maybe<Scalars['Int']>;
};
// ...
export type LastFmQuery = {
  __typename?: 'LastFMQuery';
  /** A query for chart data. */
  chart: LastFmChartQuery;
};